Added anti-spam check
Added a timer to the booking system. The user can not create another booking for 30 seconds after clicking the "Confirm Booking" button. This time is controlled by the RECENT_BOOKING_THRESHOLD_SECONDS variable. The user is presented with a warning when confirming a booking too soon.
Added an auto-clear for this warning.